Program/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
httpx>=0.27
googlemaps>=4.10
python-dotenv>=1.0
requests>=2.31
tenacity>=8.2
